+++ ios/Runner/Info.plist
@@ -36,7 +36,7 @@
 		</dict>
 	</array>
 	<key>CFBundleVersion</key>
-	<string>7</string>
+	<string>9</string>
 	<key>GADApplicationIdentifier</key>
 	<string>ca-app-pub-5619947536545679~7068085893</string>
 	<key>ITSAppUsesNonExemptEncryption</key>
