+++ lib/app/pages/bus_main_detail/binding/bus_seoul_detail_binding.dart
@@ -1,0 +1,11 @@
+import 'package:get/get.dart';
+import 'package:skkumap/app/pages/bus_main_detail/controller/bus_seoul_detail_controller.dart';
+
+class BusDetailBinding extends Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut<SeoulDetailController>(() {
+      return SeoulDetailController();
+    });
+  }
+}
