+++ lib/app/pages/mainpage/ui/mainpage_screen.dart
@@ -84,7 +84,7 @@ Widget _getSnappingSheetContent(int index) {
     case 0:
       return OptionBus();
     case 1:
-      return const OptionStation(); // Replace with your actual widget for index 1
+      return OptionStation(); // Replace with your actual widget for index 1
     case 2:
       return const OptionCampus(); // Replace with your actual widget for index 2
     default:
