+++ lib/app/utils/api_fetch/search_option3.dart
@@ -5,7 +5,9 @@ import 'package:skkumap/app/model/search_option3_model.dart';
 Future<SearchOption3Model> searchOption3(String queryString) async {
   String encodedQuery = Uri.encodeComponent(queryString);
   print("encodedQuery: $encodedQuery");
-  String url = 'http://localhost:3000/search/option3/$encodedQuery';
+  // String url = 'http://localhost:3000/search/option3/$encodedQuery';
+  String url =
+      'http://ec2-13-209-48-107.ap-northeast-2.compute.amazonaws.com/search/option3/$encodedQuery';
   // String url = 'http://ec2-13-209-48-107.ap-northeast-2.compute.amazonaws.com/mobile/v1/mainpage/buslist/$encodedQuery';
 
   final response = await http.get(Uri.parse(url));
