+++ macos/Flutter/ephemeral/Flutter-Generated.xcconfig
@@ -4,7 +4,7 @@ FLUTTER_APPLICATION_PATH=/Users/zoyoong/project/skkumap
 COCOAPODS_PARALLEL_CODE_SIGN=true
 FLUTTER_BUILD_DIR=build
 FLUTTER_BUILD_NAME=3.0.0
-FLUTTER_BUILD_NUMBER=3
+FLUTTER_BUILD_NUMBER=4
 DART_OBFUSCATION=false
 TRACK_WIDGET_CREATION=true
 TREE_SHAKE_ICONS=false
