+++ macos/Flutter/ephemeral/flutter_export_environment.sh
@@ -5,7 +5,7 @@ export "FLUTTER_APPLICATION_PATH=/Users/zoyoong/project/skkumap"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_BUILD_DIR=build"
 export "FLUTTER_BUILD_NAME=3.0.0"
-export "FLUTTER_BUILD_NUMBER=3"
+export "FLUTTER_BUILD_NUMBER=4"
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=true"
 export "TREE_SHAKE_ICONS=false"
